migrate to installable
